wip: made a `style::animation` module and moved the `Hover` struct to it
