+++ style/src/button.rs
@@ -1,7 +1,8 @@
 //! Change the apperance of a button.
-use std::time::Instant;
 
 use iced_core::{Background, Color, Vector};
+
+use crate::animation::Hover;
 
 /// The appearance of a button.
 #[derive(Debug, Clone, Copy)]
@@ -31,29 +32,6 @@ impl std::default::Default for Appearance {
             text_color: Color::BLACK,
         }
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Default)]
-/// Direction of the animation
-pub enum AnimationDirection {
-    #[default]
-    /// The animation goes forward
-    Forward,
-    /// The animation goes backward
-    Backward,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-/// Hover animation
-pub struct Hover {
-    /// Animation direction: forward means it goes from non-hovered to hovered state
-    pub direction: AnimationDirection,
-    /// The instant the animation was started at
-    pub started_at: Instant,
-    /// The progress of the animationn, between 0.0 and 1.0
-    pub animation_progress: f32,
-    /// The progress the animation has been started at
-    pub initial_progress: f32,
 }
 
 /// A set of rules that dictate the style of a button.
